test(ui): split the service lifecycle tests into their own file
Start, stop, forced override, and dependency-gated behaviour are verified as one
group, including the readiness-wait fixture they share.

Tests move unchanged.
